Fix 2 (test)

+++ src/protocolgame.h
@@ -90,7 +90,7 @@ public:
 
 	void sendLockerItems(std::map<uint16_t, uint16_t> itemMap, uint16_t count);
 
-	uint16_t getVersion() const
+	uint32_t getVersion() const
 	{
 		return version;
 	}
@@ -448,7 +448,7 @@ private:
 
 	uint32_t eventConnect = 0;
 	uint32_t challengeTimestamp = 0;
-	uint16_t version = g_config.getNumber(ConfigManager::CLIENT_VERSION);
+	uint32_t version = g_config.getNumber(ConfigManager::CLIENT_VERSION);
 	int32_t clientVersion = 0;
 
 	uint8_t challengeRandom = 0;
